`DownloadSimple` to perform artifact downloads without checking hashes

+++ src/main/java/net/runelite/launcher/SplashScreen.java
@@ -208,7 +208,7 @@ public class SplashScreen extends JFrame implements ActionListener
 
 	public static void stage(double startProgress, double endProgress,
 		@Nullable String actionText, String subActionText,
-		int done, int total, boolean mib)
+		long done, long total, boolean mib)
 	{
 		String progress;
 		if (mib)
